requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
fer==0.3.0
python-dotenv==1.0.0
requests==2.31.0
streamlit-lottie==0.0.4
openai==1.3.0
numpy==1.24.3
Pillow==10.0.1
tensorflow==2.13.0
mtcnn==0.1.1